Refactor code base into own packages
The config package now contains the structs that are used to break down
the arguments.

The get package performs all getter tasks.

The validate package simply performs the lookup and match.
