+++ env-permissions-check/permissions.go
@@ -6,50 +6,17 @@
 package main
 
 import (
-	"bufio"
 	"context"
 	"log"
 	"os"
-	"strings"
 
 	"rbac-check/pkg/client"
+	"rbac-check/pkg/config"
+	"rbac-check/pkg/get"
+	"rbac-check/pkg/validate"
 
-	"github.com/google/go-github/v35/github"
 	ghaction "github.com/sethvargo/go-githubactions"
-	"gopkg.in/yaml.v2"
 )
-
-// Rbac type is used to parse a yaml file and a list of subjects in a RoleBinding.
-type Rbac struct {
-	Subjects []Subjects `yaml:"subjects"`
-}
-
-// Subjects type is a child of Rbac and contains the name of the GitHub team.
-type Subjects struct {
-	Name string `yaml:"name"`
-}
-
-// Options defines the contents of a github client, context and other variables.
-type Options struct {
-	Client *github.Client
-	Ctx    context.Context
-	// FileName defines the file name passed by the GitHub Action.
-	FileName string
-	// AdminTeam is the admin of the users repository, i.e. can PR in all namespaces.
-	AdminTeam string
-}
-
-// User defines the structure of the user of the PR.
-type User struct {
-	// Username is the creator of the PR, passed by a GitHub Action.
-	Username string
-	// Branch is the git branch the users PR lives on.
-	Branch string
-	// Repo contains the string value of the repository this is executed against.
-	Repo string
-	// Org is the Organisation the repository/Repo lives in.
-	Org string
-}
 
 func main() {
 	// Exit hard if the environment variables don't exist. The package requires
@@ -59,14 +26,14 @@ func main() {
 		log.Fatalln("you must have the GITHUB_OAUTH_TOKEN and PR_OWNER env var set.")
 	}
 
-	user := User{
+	user := config.User{
 		Username: os.Getenv("PR_OWNER"),
 		Branch:   os.Getenv("BRANCH"),
 		Repo:     "cloud-platform-environments",
 		Org:      "ministryofjustice",
 	}
 
-	opt := Options{
+	opt := config.Options{
 		Client:    client.GitHubClient(os.Getenv("GITHUB_OAUTH_TOKEN")),
 		Ctx:       context.Background(),
 		FileName:  "files",
@@ -85,7 +52,7 @@ func main() {
 
 	// Parse all namespaces in the fileName variable. Fail if we can't parse because later
 	// functions require this output.
-	namespaces, err := getNamespaces(opt.FileName)
+	namespaces, err := get.Namespaces(opt.FileName)
 	if err != nil {
 		log.Fatalln("Unable to fetch namespace:", err)
 	}
@@ -95,7 +62,7 @@ func main() {
 	// deduplication in Go.
 	namespaceTeams := make(map[string]int)
 	for ns := range namespaces {
-		teams, err := getTeamName(ns, &opt, &user)
+		teams, err := get.TeamName(ns, &opt, &user)
 		if err != nil {
 			log.Fatalln("Unable to get team names:", err)
 		}
@@ -113,13 +80,13 @@ func main() {
 
 	// Convert the PR_OWNER string into a GitHub user ID. This is used later, to compare the
 	// list of users in a team.
-	userID, err := getUserID(&opt, &user)
+	userID, err := get.UserID(&opt, &user)
 	if err != nil {
 		log.Fatalln("Unable to fetch userID", err)
 	}
 
 	// Call the GitHub API to confirm if the user exists in the GitHub team name.
-	valid, team, err := isUserValid(namespaceTeams, userID, &opt, &user)
+	valid, team, err := validate.UserPermissions(namespaceTeams, userID, &opt, &user)
 	if err != nil {
 		log.Fatalln("Unable to check if the user is valid:", err)
 	}
@@ -132,128 +99,4 @@ func main() {
 		log.Println("\n The user:", userID.GetName(), "\n can't be found in teams:", namespaceTeams)
 		ghaction.SetOutput("review_pr", "false")
 	}
-}
-
-func getUserID(opt *Options, user *User) (*github.User, error) {
-	userID, _, err := opt.Client.Users.Get(opt.Ctx, user.Username)
-	if err != nil {
-		return nil, err
-	}
-
-	return userID, nil
-}
-
-func getOrigin(namespace string, opt *Options, user *User, repoOpts *github.RepositoryContentGetOptions) (string, error) {
-	secondaryCluster := "live"
-	primaryCluster := "live-1"
-
-	cluster := primaryCluster
-	path := "namespaces/" + cluster + ".cloud-platform.service.justice.gov.uk/" + namespace + "/01-rbac.yaml"
-
-	_, _, resp, err := opt.Client.Repositories.GetContents(opt.Ctx, user.Org, user.Repo, path, repoOpts)
-	if err != nil {
-		return "", err
-	}
-
-	if resp.StatusCode == 200 {
-		return cluster, nil
-	} else {
-		cluster = secondaryCluster
-		_, _, resp, err := opt.Client.Repositories.GetContents(opt.Ctx, user.Org, user.Repo, path, repoOpts)
-		if err != nil {
-			return "", err
-		}
-		if resp.StatusCode == 200 {
-			return cluster, nil
-		}
-	}
-
-	return "none", nil
-}
-
-func getTeamName(namespace string, opt *Options, user *User) ([]string, error) {
-	repoOpts := &github.RepositoryContentGetOptions{}
-	// 3 cases here. Live-1, live if a namespace doesn't yet exist
-
-	// is it live, live-1 or doesn't it exist
-	// Find out if it's live-1, live or in the PR.
-	origin, err := getOrigin(namespace, opt, user, repoOpts)
-	if err != nil {
-		log.Println(err)
-	}
-
-	// if the namespace doesn't exist yet, check the pr.
-	if origin == "none" {
-		repoOpts = &github.RepositoryContentGetOptions{
-			Ref: user.Branch,
-		}
-		origin, err = getOrigin(namespace, opt, user, repoOpts)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	path := "namespaces/" + origin + ".cloud-platform.service.justice.gov.uk/" + namespace + "/01-rbac.yaml"
-	file, _, _, err := opt.Client.Repositories.GetContents(opt.Ctx, user.Org, user.Repo, path, repoOpts)
-	if err != nil {
-		return nil, err
-	}
-
-	cont, err := file.GetContent()
-	if err != nil {
-		return nil, err
-	}
-
-	fullName := Rbac{}
-
-	err = yaml.Unmarshal([]byte(cont), &fullName)
-	if err != nil {
-		return nil, err
-	}
-
-	var namespaceTeams []string
-	for _, name := range fullName.Subjects {
-		str := strings.SplitAfter(string(name.Name), ":")
-		namespaceTeams = append(namespaceTeams, str[1])
-	}
-
-	return namespaceTeams, nil
-}
-
-func getNamespaces(fileName string) (map[string]int, error) {
-	namespaces := make(map[string]int)
-
-	file, err := os.Open(fileName)
-	if err != nil {
-		return namespaces, err
-	}
-	scanner := bufio.NewScanner(file)
-	scanner.Split(bufio.ScanWords)
-
-	for scanner.Scan() {
-		if namespaces[scanner.Text()] == 0 {
-			namespaces[scanner.Text()] = 1
-		} else {
-			namespaces[scanner.Text()]++
-		}
-	}
-
-	return namespaces, nil
-}
-
-func isUserValid(namespaceTeams map[string]int, githubUser *github.User, opt *Options, user *User) (bool, string, error) {
-	teamOpts := &github.TeamListTeamMembersOptions{}
-	for team := range namespaceTeams {
-		members, _, err := opt.Client.Teams.ListTeamMembersBySlug(opt.Ctx, user.Org, team, teamOpts)
-		if err != nil {
-			return false, "", nil
-		}
-		for _, member := range members {
-			if member.GetID() == githubUser.GetID() {
-				return true, team, nil
-			}
-		}
-	}
-
-	return false, "", nil
 }
